Update index.js
Sarai Hannah Ajai have added a middleware function that checks, if the user is already authenticated. If the user is not authenticated then the middleware sets the req.isUnauthenticated flag to true. This middleware function is added all to the routes using app.use method.

In all the routes, Sarai Hannah Ajai checks, if the user is already authenticated. if the user is authenticated then the res.render route is release or a message is release letting the user know he/she is already logged in. If the user is not authenticated then route is check, if this is the first time use of the user route, '/signup' in the URL by ways of checking the req.isUnauthenticated flag. If this is, the first time user use in the URL session then the signup web page is render to the user dated 03-10-23.